feat(agents): cleans agents header ab test
