+++ types.ts
@@ -3,7 +3,6 @@ export type Entries<T> = {
 }[keyof T][];
 
 export interface MatomoAbTestVariations {
-  AgentHeader: "original" | "VariationA";
   AgentWall: "original" | "VariationA" | "VariationB";
 }
